+++ pipelines/core/constants.py
