+++ pipelines/core/settings.py
