+++ pipelines/tilesets/cli.py
@@ -7,10 +7,10 @@ from urllib.parse import urlparse
 from subprocess import run
 from typing import Iterable
 from files import download_and_unzip_zipfile, exists_in_s3
-from settings import settings
+from core.settings import settings
 from models import GerryDBTileset, TilesetBatch
 from utils import merge_tilesets
-from constants import (
+from core.constants import (
     DEFAULT_GERRYDB_COLUMNS,
     S3_BASEMAPS_PREFIX,
     TIGER_COUNTY_URL,
